Move toolbar and plugin config
