+++ src/components/feature-editor/config.ts
@@ -1,0 +1,101 @@
+const boardToolbarSimple = [
+	"heading",
+	"|",
+	"bold",
+	"italic",
+	"highlight",
+	"|",
+	"link",
+	"bulletedList",
+	"numberedList",
+	"removeFormat",
+];
+
+const boardToolbarRegular = [
+	"undo",
+	"redo",
+	"|",
+	"heading",
+	"|",
+	"bold",
+	"italic",
+	"underline",
+	"strikethrough",
+	"highlight",
+	"fontBackgroundColor",
+	"code",
+	"superscript",
+	"subscript",
+	"|",
+	"link",
+	"bulletedList",
+	"numberedList",
+	"math",
+	"horizontalLine",
+	"|",
+	"blockQuote",
+	"insertTable",
+	"specialCharacters",
+	"removeFormat",
+];
+
+const newsToolbar = [
+	"undo",
+	"redo",
+	"|",
+	"heading",
+	"|",
+	"bold",
+	"italic",
+	"insertImage",
+	"bulletedList",
+	"numberedList",
+	"|",
+	"removeFormat",
+];
+
+const boardPlugins = [
+	"Autoformat",
+	"Essentials",
+	"BlockQuote",
+	"Bold",
+	"Code",
+	"Heading",
+	"Highlight",
+	"HorizontalLine",
+	"Italic",
+	"Link",
+	"List",
+	"Math",
+	"Paragraph",
+	"RemoveFormat",
+	"SpecialCharacters",
+	"Strikethrough",
+	"Subscript",
+	"Superscript",
+	"Table",
+	"TableToolbar",
+	"WordCount",
+];
+
+const newsPlugins = [
+	"Autoformat",
+	"Essentials",
+	"Bold",
+	"Heading",
+	"Italic",
+	"List",
+	"Paragraph",
+	"RemoveFormat",
+	"Strikethrough",
+	"Image",
+	"ImageInsertViaUrl",
+];
+
+export {
+	boardToolbarSimple,
+	boardToolbarRegular,
+	newsToolbar,
+	boardPlugins,
+	newsPlugins,
+};
